Refactor connection creation
Remove is_auth method. Move connect_to_snowflake to snow_connector module.
Add nicer handling of Snowflake exceptions when estabilishing connection.
