+++ src/snowcli/cli/snowpark/cp.py
@@ -3,7 +3,7 @@ from snowcli import config
 from snowcli.cli import DEFAULT_CONTEXT_SETTINGS
 from snowcli.cli.common.alias import build_alias
 from snowcli.cli.common.flags import ConnectionOption
-from snowcli.config import connect_to_snowflake
+from snowcli.snow_connector import connect_to_snowflake
 from snowcli.output.printing import print_db_cursor
 
 app = typer.Typer(
@@ -25,17 +25,16 @@ def create(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.create_compute_pool(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-            num_instances=num_instances,
-            instance_family=instance_family,
-        )
-        print_db_cursor(results)
+    results = conn.create_compute_pool(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+        num_instances=num_instances,
+        instance_family=instance_family,
+    )
+    print_db_cursor(results)
 
 
 @app.command()
@@ -45,14 +44,13 @@ def list(environment: str = ConnectionOption):
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.list_compute_pools(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-        )
-        print_db_cursor(results)
+    results = conn.list_compute_pools(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+    )
+    print_db_cursor(results)
 
 
 @app.command()
@@ -65,15 +63,14 @@ def drop(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.drop_compute_pool(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-        )
-        print_db_cursor(results)
+    results = conn.drop_compute_pool(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+    )
+    print_db_cursor(results)
 
 
 @app.command()
@@ -86,15 +83,14 @@ def stop(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.stop_compute_pool(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-        )
-        print_db_cursor(results)
+    results = conn.stop_compute_pool(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+    )
+    print_db_cursor(results)
 
 
 app_cp = build_alias(
