+++ src/snowcli/cli/snowpark/procedure_coverage/report.py
@@ -8,7 +8,7 @@ import snowflake.connector
 import typer
 
 from snowcli import config, utils
-from snowcli.config import connect_to_snowflake
+from snowcli.snow_connector import connect_to_snowflake
 from snowcli.utils import generate_deploy_stage_name
 
 from snowcli.cli.snowpark.procedure_coverage import app
@@ -54,61 +54,60 @@ def procedure_coverage_report(
     ),
 ):
     conn = connect_to_snowflake(connection_name=environment)
-    if config.is_auth():
-        deploy_dict = utils.get_deploy_names(
-            conn.ctx.database,
-            conn.ctx.schema,
-            generate_deploy_stage_name(
-                name,
-                input_parameters,
-            ),
-        )
-        coverage_file = ".coverage"
-        # the coverage databases will include paths like "/home/udf/132735964617982/app.zip/app.py", where they ran on Snowflake
-        # we need to strip out the prefix up to and including "app.zip/" so that it reads them from the local folder
-        # tried to do this by modifying the report data, but couldn't seem to figure it out
-        # instead we'll monkey-patch the source code reader and strip it out
-        orig_get_python_source = coverage.python.get_python_source
+    deploy_dict = utils.get_deploy_names(
+        conn.ctx.database,
+        conn.ctx.schema,
+        generate_deploy_stage_name(
+            name,
+            input_parameters,
+        ),
+    )
+    coverage_file = ".coverage"
+    # the coverage databases will include paths like "/home/udf/132735964617982/app.zip/app.py", where they ran on Snowflake
+    # we need to strip out the prefix up to and including "app.zip/" so that it reads them from the local folder
+    # tried to do this by modifying the report data, but couldn't seem to figure it out
+    # instead we'll monkey-patch the source code reader and strip it out
+    orig_get_python_source = coverage.python.get_python_source
 
-        def new_get_python_source(filename: str):
-            new_filename = filename[filename.index("app.zip/") + len("app.zip/") :]
-            return orig_get_python_source(new_filename)
+    def new_get_python_source(filename: str):
+        new_filename = filename[filename.index("app.zip/") + len("app.zip/") :]
+        return orig_get_python_source(new_filename)
 
-        coverage.python.get_python_source = new_get_python_source
+    coverage.python.get_python_source = new_get_python_source
 
-        combined_coverage = coverage.Coverage(data_file=coverage_file)
-        with tempfile.TemporaryDirectory() as temp_dir:
-            stage_name = deploy_dict["stage"]
-            stage_path = deploy_dict["directory"]
-            report_files = f"{stage_name}{stage_path}/coverage/"
-            try:
-                results = conn.get_stage(
-                    database=conn.ctx.database,
-                    schema=conn.ctx.schema,
-                    role=conn.ctx.role,
-                    warehouse=conn.ctx.warehouse,
-                    name=report_files,
-                    path=str(temp_dir),
-                ).fetchall()
-            except snowflake.connector.errors.DatabaseError as database_error:
-                if database_error.errno == 253006:
-                    results = []
-            if len(results) == 0:
-                log.error(
-                    "No code coverage reports were found on the stage. "
-                    "Please ensure that you've invoked the procedure at least once "
-                    "and that you provided the correct inputs"
-                )
-                raise typer.Abort()
-            else:
-                log.info(f"Combining data from {len(results)} reports")
-            combined_coverage.combine(
-                # the tuple contains the columns: | file ┃ size ┃ status ┃ message |
-                data_paths=[
-                    os.path.join(temp_dir, os.path.basename(result[0]))
-                    for result in results
-                ]
+    combined_coverage = coverage.Coverage(data_file=coverage_file)
+    with tempfile.TemporaryDirectory() as temp_dir:
+        stage_name = deploy_dict["stage"]
+        stage_path = deploy_dict["directory"]
+        report_files = f"{stage_name}{stage_path}/coverage/"
+        try:
+            results = conn.get_stage(
+                database=conn.ctx.database,
+                schema=conn.ctx.schema,
+                role=conn.ctx.role,
+                warehouse=conn.ctx.warehouse,
+                name=report_files,
+                path=str(temp_dir),
+            ).fetchall()
+        except snowflake.connector.errors.DatabaseError as database_error:
+            if database_error.errno == 253006:
+                results = []
+        if len(results) == 0:
+            log.error(
+                "No code coverage reports were found on the stage. "
+                "Please ensure that you've invoked the procedure at least once "
+                "and that you provided the correct inputs"
             )
+            raise typer.Abort()
+        else:
+            log.info(f"Combining data from {len(results)} reports")
+        combined_coverage.combine(
+            # the tuple contains the columns: | file ┃ size ┃ status ┃ message |
+            data_paths=[
+                os.path.join(temp_dir, os.path.basename(result[0]))
+                for result in results
+            ]
+        )
         if output_format == ReportOutputOptions.html:
             coverage_percentage = combined_coverage.html_report()
             log.info(
