+++ src/snowcli/cli/stage.py
@@ -4,9 +4,10 @@ from pathlib import Path
 
 import typer
 
+import snowcli.snow_connector
 from snowcli import config
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS, ConnectionOption
-from snowcli.config import connect_to_snowflake
+from snowcli.snow_connector import connect_to_snowflake
 from snowcli.output.printing import print_db_cursor
 
 app = typer.Typer(
@@ -26,24 +27,23 @@ def stage_list(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        if name:
-            results = conn.list_stage(
-                database=conn.ctx.database,
-                schema=conn.ctx.schema,
-                role=conn.ctx.role,
-                warehouse=conn.ctx.warehouse,
-                name=name,
-            )
-            print_db_cursor(results)
-        else:
-            results = conn.list_stages(
-                database=conn.ctx.database,
-                schema=conn.ctx.schema,
-                role=conn.ctx.role,
-                warehouse=conn.ctx.warehouse,
-            )
-            print_db_cursor(results)
+    if name:
+        results = conn.list_stage(
+            database=conn.ctx.database,
+            schema=conn.ctx.schema,
+            role=conn.ctx.role,
+            warehouse=conn.ctx.warehouse,
+            name=name,
+        )
+        print_db_cursor(results)
+    else:
+        results = conn.list_stages(
+            database=conn.ctx.database,
+            schema=conn.ctx.schema,
+            role=conn.ctx.role,
+            warehouse=conn.ctx.warehouse,
+        )
+        print_db_cursor(results)
 
 
 @app.command("get")
@@ -65,16 +65,15 @@ def stage_get(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.get_stage(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-            path=str(path),
-        )
-        print_db_cursor(results)
+    results = conn.get_stage(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+        path=str(path),
+    )
+    print_db_cursor(results)
 
 
 @app.command("put")
@@ -104,22 +103,21 @@ def stage_put(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        filepath = str(path)
-        if path.is_dir():
-            filepath = str(path) + "/*"
+    filepath = str(path)
+    if path.is_dir():
+        filepath = str(path) + "/*"
 
-        results = conn.put_stage(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-            path=str(filepath),
-            overwrite=overwrite,
-            parallel=parallel,
-        )
-        print_db_cursor(results)
+    results = conn.put_stage(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+        path=str(filepath),
+        overwrite=overwrite,
+        parallel=parallel,
+    )
+    print_db_cursor(results)
 
 
 @app.command("create")
@@ -132,15 +130,14 @@ def stage_create(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.create_stage(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-        )
-        print_db_cursor(results)
+    results = conn.create_stage(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+    )
+    print_db_cursor(results)
 
 
 @app.command("drop")
@@ -153,15 +150,14 @@ def stage_drop(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.drop_stage(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-        )
-        print_db_cursor(results)
+    results = conn.drop_stage(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+    )
+    print_db_cursor(results)
 
 
 @app.command("remove")
@@ -176,14 +172,13 @@ def stage_remove(
 
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        config.connect_to_snowflake()
-        results = conn.remove_from_stage(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=stage_name,
-            path=file_name,
-        )
-        print_db_cursor(results)
+    snowcli.snow_connector.connect_to_snowflake()
+    results = conn.remove_from_stage(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=stage_name,
+        path=file_name,
+    )
+    print_db_cursor(results)
