+++ src/snowcli/cli/streamlit.py
@@ -7,7 +7,7 @@ from typing import List, Optional
 
 from snowcli import config
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS, ConnectionOption
-from snowcli.config import connect_to_snowflake
+from snowcli.snow_connector import connect_to_snowflake
 from snowcli.output.printing import print_db_cursor
 from snowcli.utils import (
     generate_streamlit_environment_file,
@@ -45,18 +45,17 @@ def streamlit_list(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.list_streamlits(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-        )
+    results = conn.list_streamlits(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+    )
 
-        print_db_cursor(
-            results,
-            columns=only_cols,
-        )
+    print_db_cursor(
+        results,
+        columns=only_cols,
+    )
 
 
 @app.command("describe")
@@ -69,16 +68,15 @@ def streamlit_describe(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        description, url = conn.describe_streamlit(
-            name,
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-        )
-        print_db_cursor(description)
-        print_db_cursor(url)
+    description, url = conn.describe_streamlit(
+        name,
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+    )
+    print_db_cursor(description)
+    print_db_cursor(url)
 
 
 @app.command("create")
@@ -107,27 +105,26 @@ def streamlit_create(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        if from_stage:
-            if "." in from_stage:
-                full_stage_name = from_stage
-            else:
-                full_stage_name = f"{conn.ctx.database}.{conn.ctx.schema}.{from_stage}"
-            standard_page_name = get_standard_stage_name(full_stage_name)
-            from_stage_command = f"FROM {standard_page_name}"
+    if from_stage:
+        if "." in from_stage:
+            full_stage_name = from_stage
         else:
-            from_stage_command = ""
+            full_stage_name = f"{conn.ctx.database}.{conn.ctx.schema}.{from_stage}"
+        standard_page_name = get_standard_stage_name(full_stage_name)
+        from_stage_command = f"FROM {standard_page_name}"
+    else:
+        from_stage_command = ""
 
-        results = conn.create_streamlit(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-            file="streamlit_app_launcher.py" if use_packaging_workaround else file.name,
-            from_stage_command=from_stage_command,
-        )
-        print_db_cursor(results)
+    results = conn.create_streamlit(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+        file="streamlit_app_launcher.py" if use_packaging_workaround else file.name,
+        from_stage_command=from_stage_command,
+    )
+    print_db_cursor(results)
 
 
 @app.command("share")
@@ -143,16 +140,15 @@ def streamlit_share(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.share_streamlit(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-            to_role=to_role,
-        )
-        print_db_cursor(results)
+    results = conn.share_streamlit(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+        to_role=to_role,
+    )
+    print_db_cursor(results)
 
 
 @app.command("drop")
@@ -165,15 +161,14 @@ def streamlit_drop(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.drop_streamlit(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-            name=name,
-        )
-        print_db_cursor(results)
+    results = conn.drop_streamlit(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+        name=name,
+    )
+    print_db_cursor(results)
 
 
 @app.command("deploy")
@@ -219,108 +214,107 @@ def streamlit_deploy(
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        schema = conn.ctx.schema
-        role = conn.ctx.role
-        database = conn.ctx.database
-        warehouse = conn.ctx.warehouse
-        # THIS WORKAROUND HAS NOT BEEN TESTETD WITH THE NEW STREAMLIT SYNTAX
-        if use_packaging_workaround:
-            # package an app.zip file, same as the other snowpark_containers_cmds package commands
-            snowpark_package(
-                pypi_download,  # type: ignore[arg-type]
-                check_anaconda_for_pypi_deps,
-                package_native_libraries,  # type: ignore[arg-type]
-            )
-            # upload the resulting app.zip file
-            conn.upload_file_to_stage(
-                "app.zip",
-                f"{name}_stage",
-                "/",
-                role=role,
-                database=database,
-                schema=schema,
-                warehouse=warehouse,
-                overwrite=True,
-                create_stage=False,
-            )
-            main_module = str(file).replace(".py", "")
-            file = generate_streamlit_package_wrapper(
-                stage_name=f"{name}_stage",
-                main_module=main_module,
-                extract_zip=packaging_workaround_includes_content,
-            )
-            # upload the wrapper file
-            conn.upload_file_to_stage(
-                str(file),
-                f"{name}_stage",
-                "/",
-                role=role,
-                database=database,
-                schema=schema,
-                warehouse=warehouse,
-                overwrite=True,
-                create_stage=False,
-            )
-            # if the packaging process generated an environment.snowflake.txt
-            # file, convert it into an environment.yml file
-            excluded_anaconda_deps_list: Optional[List[str]] = None
-            if excluded_anaconda_deps is not None:
-                excluded_anaconda_deps_list = excluded_anaconda_deps.split(",")
-            env_file = generate_streamlit_environment_file(excluded_anaconda_deps_list)
-            if env_file:
-                conn.upload_file_to_stage(
-                    str(env_file),
-                    f"{name}_stage",
-                    "/",
-                    role=role,
-                    database=database,
-                    schema=schema,
-                    warehouse=warehouse,
-                    overwrite=True,
-                    create_stage=False,
-                )
-
-        base_url = conn.deploy_streamlit(
-            name=name,
-            file_path=str(file),
-            stage_path="/",
+    schema = conn.ctx.schema
+    role = conn.ctx.role
+    database = conn.ctx.database
+    warehouse = conn.ctx.warehouse
+    # THIS WORKAROUND HAS NOT BEEN TESTETD WITH THE NEW STREAMLIT SYNTAX
+    if use_packaging_workaround:
+        # package an app.zip file, same as the other snowpark_containers_cmds package commands
+        snowpark_package(
+            pypi_download,  # type: ignore[arg-type]
+            check_anaconda_for_pypi_deps,
+            package_native_libraries,  # type: ignore[arg-type]
+        )
+        # upload the resulting app.zip file
+        conn.upload_file_to_stage(
+            "app.zip",
+            f"{name}_stage",
+            "/",
             role=role,
             database=database,
             schema=schema,
             warehouse=warehouse,
             overwrite=True,
+            create_stage=False,
         )
-
-        def get_url() -> str:
-            try:
-                host = conn.ctx.host
-            except KeyError:
-                return base_url
-
-            host_parts = host.split(".")
-
-            if len(host_parts) != 6:
-                log.error(
-                    f"The connection host ({host}) was missing or not in "
-                    "the expected format "
-                    "(<account>.<deployment>.snowflakecomputing.com)"
-                )
-                raise typer.Exit()
-            else:
-                account_name = host_parts[0]
-                deployment = ".".join(host_parts[1:4])
-
-            snowflake_host = conn.ctx.host or "app.snowflake.com"
-            uppercased_dsn = f"{database}.{schema}.{name}".upper()
-            return (
-                f"https://{snowflake_host}/{deployment}/{account_name}/"
-                f"#/streamlit-apps/{uppercased_dsn}"
+        main_module = str(file).replace(".py", "")
+        file = generate_streamlit_package_wrapper(
+            stage_name=f"{name}_stage",
+            main_module=main_module,
+            extract_zip=packaging_workaround_includes_content,
+        )
+        # upload the wrapper file
+        conn.upload_file_to_stage(
+            str(file),
+            f"{name}_stage",
+            "/",
+            role=role,
+            database=database,
+            schema=schema,
+            warehouse=warehouse,
+            overwrite=True,
+            create_stage=False,
+        )
+        # if the packaging process generated an environment.snowflake.txt
+        # file, convert it into an environment.yml file
+        excluded_anaconda_deps_list: Optional[List[str]] = None
+        if excluded_anaconda_deps is not None:
+            excluded_anaconda_deps_list = excluded_anaconda_deps.split(",")
+        env_file = generate_streamlit_environment_file(excluded_anaconda_deps_list)
+        if env_file:
+            conn.upload_file_to_stage(
+                str(env_file),
+                f"{name}_stage",
+                "/",
+                role=role,
+                database=database,
+                schema=schema,
+                warehouse=warehouse,
+                overwrite=True,
+                create_stage=False,
             )
 
-        url = get_url()
+    base_url = conn.deploy_streamlit(
+        name=name,
+        file_path=str(file),
+        stage_path="/",
+        role=role,
+        database=database,
+        schema=schema,
+        warehouse=warehouse,
+        overwrite=True,
+    )
 
-        if open_:
-            typer.launch(url)
+    def get_url() -> str:
+        try:
+            host = conn.ctx.host
+        except KeyError:
+            return base_url
+
+        host_parts = host.split(".")
+
+        if len(host_parts) != 6:
+            log.error(
+                f"The connection host ({host}) was missing or not in "
+                "the expected format "
+                "(<account>.<deployment>.snowflakecomputing.com)"
+            )
+            raise typer.Exit()
         else:
-            log.info(url)
+            account_name = host_parts[0]
+            deployment = ".".join(host_parts[1:4])
+
+        snowflake_host = conn.ctx.host or "app.snowflake.com"
+        uppercased_dsn = f"{database}.{schema}.{name}".upper()
+        return (
+            f"https://{snowflake_host}/{deployment}/{account_name}/"
+            f"#/streamlit-apps/{uppercased_dsn}"
+        )
+
+    url = get_url()
+
+    if open_:
+        typer.launch(url)
+    else:
+        log.info(url)
