+++ src/snowcli/cli/warehouse.py
@@ -4,7 +4,7 @@ import typer
 
 from snowcli import config
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS, ConnectionOption
-from snowcli.config import connect_to_snowflake
+from snowcli.snow_connector import connect_to_snowflake
 from snowcli.output.printing import print_db_cursor
 
 app = typer.Typer(
@@ -22,11 +22,10 @@ def warehouse_status(environment: str = ConnectionOption):
     """
     conn = connect_to_snowflake(connection_name=environment)
 
-    if config.is_auth():
-        results = conn.show_warehouses(
-            database=conn.ctx.database,
-            schema=conn.ctx.schema,
-            role=conn.ctx.role,
-            warehouse=conn.ctx.warehouse,
-        )
-        print_db_cursor(results, ["name", "state", "queued", "resumed_on"])
+    results = conn.show_warehouses(
+        database=conn.ctx.database,
+        schema=conn.ctx.schema,
+        role=conn.ctx.role,
+        warehouse=conn.ctx.warehouse,
+    )
+    print_db_cursor(results, ["name", "state", "queued", "resumed_on"])
